refactor: shared resource between command and testing
